add setup for mac

+++ setup/setup.ps1
@@ -1,4 +1,8 @@
 # Window Setup
+$env:OS = "Windows"
+Write-Host "Environment variable set: OS=$($env:OS)"
+
+# -- Setup Dotfiles --
 
 $dotfiles = Join-Path $home "dotfiles"
 
@@ -10,19 +14,68 @@ $symlinks = @(
     @{ Source = "dist\windows\Microsoft.PowerShell_profile.ps1"; Destination = $profile }
 )
 
+$isAdmin = ([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
 
-foreach ($link in $symlinks) {
-    $sourcePath = Join-Path $dotfiles $link.Source
-    $destinationPath = $link.Destination
-	
-    # Remove existing config if it exists
-    if (Test-Path $destinationPath) {
-        Remove-Item $destinationPath -Recurse -Force
-	    Write-Host "Removed existing: $destinationPath"
+if ($isAdmin) {
+    foreach ($link in $symlinks) {
+        $sourcePath = Join-Path $dotfiles $link.Source
+        $destinationPath = $link.Destination
+	    
+        # Remove existing config if it exists
+        if (Test-Path $destinationPath) {
+            Remove-Item $destinationPath -Recurse -Force
+	        Write-Host "Removed existing: $destinationPath"
+        }
+
+    # Create symlink
+    New-Item -ItemType SymbolicLink -Path $destinationPath -Target $sourcePath
+    Write-Host "Symlink created: $destinationPath -> $sourcePath"
     }
+} else {
+    Write-Host "Detected not in administrator mode, symlinks skipped"
+}
 
-   # Create symlink
-   New-Item -ItemType SymbolicLink -Path $destinationPath -Target $sourcePath
-   Write-Host "Symlink created: $destinationPath -> $sourcePath"
+# -- Install Dependencies --
+$wingetApps = @(
+    "Neovim.Neovim",
+    "sharkdp.fd",
+    "BurntSushi.ripgrep.MSVC",
+    "JesseDuffield.lazygit",
+    "wez.wezterm",
+    "Starship.Starship",
+    "ajeetdsouza.zoxide",
+    "DEVCOM.JetBrainsMonoNerdFont"
+    "fzf"
+
+    # Dist Specific
+    "GlazeWM"
+)
+
+foreach ($appId in $wingetApps) {
+    $installed = winget list --id=$appId -e | Select-String $appId
+
+    if ($installed) {
+        Write-Host "$appId already installed. Skipping..."
+    } else {
+        Write-Host "Installing $appId"
+        winget install $appId
+    }
+}
+
+$modules = @(
+    "PowerShellGet",
+    "PSReadLine",
+    "PSFzf"
+)
+
+foreach ($module in $modules) {
+    $installed = Get-Module -ListAvailable -Name $module
+
+    if ($installed) {
+        Write-Host "$module already installed. Skipping..."
+    } else {
+        Write-Host "Installing $module"
+        Install-Module -Name $module -Force
+    }
 }
